refactor(pages/components/helper/togglerow): separacao de funcoes
@janinelps @eluvya @Cibely87

78
