+++ src/pages/challenges/components/toggle-row/index.js
@@ -7,102 +7,16 @@ import { client } from '../../../../service'
 import { useNavigate } from 'react-router-dom'
 import { Tr } from './styled'
 import { useTranslation } from 'react-i18next'
-
-const statusEnum = {
-  PREPARING: 'status-preparing',
-  CLOSED: 'status-closed',
-  OPENED: 'status-opened'
-}
-
-const hasFeedback = (challenge) => {
-  if (!challenge.exercises) {
-    return false
-  }
-  return challenge.exercises.find((feed) => feed.feedback !== null)
-}
-
-const hasMentorName = (challenge) => {
-  if (!challenge.exercises) {
-    return false
-  }
-  return challenge.exercises.find((nameMentor) => nameMentor.mentorName !== null)
-}
-
-const wasCanceled = (challenge) => {
-  if (!challenge.exercises) {
-    return false
-  }
-  return challenge.exercises.find((name) => name.mentorName === 'cancelado')
-}
-
-const getStatus = (challenge) => {
-  if (!challenge) return statusEnum.OPENED
-  if (wasCanceled(challenge)) return statusEnum.OPENED
-  if (hasFeedback(challenge) && hasMentorName(challenge)) return statusEnum.CLOSED
-  if (hasMentorName(challenge)) return statusEnum.PREPARING
-  return statusEnum.OPENED
-}
-
-const getStatusEvaluation = (evaluation) => {
-  if (!evaluation) return statusEnum.OPENED
-  const { feedback, mentorName } = evaluation
-  if (mentorName === 'cancelado') return statusEnum.OPENED
-  if (feedback && mentorName) return statusEnum.CLOSED
-  if (mentorName) return statusEnum.PREPARING
-  return statusEnum.OPENED
-}
-
-const isClosedChallenge = (exercises) => {
-  return exercises.filter((ex) => getStatusEvaluation(ex.evaluation) !== statusEnum.CLOSED).length === 0
-}
-
-const isPreparedChallenge = (exercises, mentorNameLocal) => {
-  return exercises.some((ex) => {
-    return isPreparing({
-      status: getStatusEvaluation(ex.evaluation),
-      currentMentor: ex.evaluation.mentorName || '',
-      actualMentor: mentorNameLocal
-    }) === true
-  })
-}
-
-const isPrepared = ({ status }) => status === statusEnum.PREPARING
-
-const isOpened = ({ status }) => status === statusEnum.OPENED
-
-const isClosed = ({ status }) => status === statusEnum.CLOSED
-
-const isPreparingOrOpened = ({ status }) => isPrepared({ status }) || isOpened({ status })
-
-const isClosedOrPreparing = ({ status }) => isPrepared({ status }) || isClosed({ status })
-
-const isPreparing = ({ status, currentMentor, actualMentor }) =>
-  isPrepared({ status }) && currentMentor !== actualMentor
-
-const getNumberChallenge = (exercises) => {
-  const statusEvaluation = exercises.map(exercise => getStatusEvaluation(exercise.evaluation))
-  const closed = statusEvaluation.filter(status => status === statusEnum.CLOSED).length
-  return { closed, total: statusEvaluation.length }
-}
-
-const getFeedback = (exercises, toggle) => {
-  const feedbacks = exercises.filter(exercise => exercise.evaluation.feedback)
-  return (
-    <tr>
-      <td colSpan="6" className={toggle}>
-        <div style={{ display: 'flex', justifyContent: 'space-between' }}>
-          {feedbacks.map((feedback, index) => (
-            <div key={index}>
-              <strong>{feedback.name}</strong>
-              <div>{feedback.evaluation.feedback} </div>
-            </div>
-          )
-          )}
-        </div>
-      </td>
-    </tr>
-  )
-}
+import {
+  getNumberChallenge,
+  getStatus,
+  getStatusEvaluation,
+  hasFeedback,
+  isClosedChallenge,
+  isClosedOrPreparing,
+  isPreparedChallenge,
+  isPreparingOrOpened
+} from '../helper'
 
 export const ToggleRow = ({ item }) => {
   const { t } = useTranslation()
@@ -168,7 +82,7 @@ export const ToggleRow = ({ item }) => {
             icon={faAngleDown} />
         }</td>
       </Tr>
-      {status && hasFeedback(item) ? getFeedback(item.exercises, toggle) : null}
+      {status && hasFeedback(item) ? <getFeedback exercises={item} toggle={item} /> : null}
     </>
   )
 }
